requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.5.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
easydict>=1.10
timm>=0.9.0  # PyTorch Image Models

# Metrics and Evaluation
torchmetrics>=1.0.0
scikit-image>=0.21.0

# Documentation
jupyter>=1.0.0
ipykernel>=6.23.0
notebook>=6.5.4

# Code Quality
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

# For PDF generation
reportlab>=4.0.0
markdown>=3.4.0
pypandoc>=1.11

# Additional utilities
einops>=0.6.1
